requirements.txt
beautifulsoup4
espn_api
pyyaml
sqlalchemy
psycopg2-binary
